Rewrote code from functional style to interfaces approach.
