+++ TagCloudGenerator/TagCloudGenerator.Tests/TestFixtures/CircularCloudLayouterTests.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using FluentAssertions;
 using NUnit.Framework;
 using NUnit.Framework.Interfaces;
-using TagCloudGenerator.CloudLayouters;
+using TagCloudGenerator.GeneratorCore.CloudLayouters;
 using TagCloudGenerator.Tests.Extensions;
 using TagCloudGenerator.Tests.WrongVisualization;
