+++ TagCloudGenerator/TagCloudGenerator/Clients/IClient.cs
@@ -1,0 +1,9 @@
+using TagCloudGenerator.GeneratorCore.TagClouds;
+
+namespace TagCloudGenerator.Clients
+{
+    public interface IClient
+    {
+        ITagCloudOptions<ITagCloud> GetOptions();
+    }
+}
