--- TagCloudGenerator/TagCloudGenerator/CloudLayouters/ICloudLayouter.cs
@@ -1,9 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagCloudGenerator.CloudLayouters
-{
-    public interface ICloudLayouter
-    {
-        Rectangle PutNextRectangle(Size rectangleSize);
-    }
-}
