--- TagCloudGenerator/TagCloudGenerator/CloudVocabularyPreprocessors/ToLowerPreprocessor.cs
@@ -1,13 +1,0 @@
-using System.Collections.Generic;
-using System.Linq;
-
-namespace TagCloudGenerator.CloudVocabularyPreprocessors
-{
-    public class ToLowerPreprocessor : CloudVocabularyPreprocessor
-    {
-        public ToLowerPreprocessor(CloudVocabularyPreprocessor nextPreprocessor) : base(nextPreprocessor) { }
-
-        protected override IEnumerable<string> ProcessVocabulary(IEnumerable<string> words) =>
-            words.Select(word => word.ToLower());
-    }
-}
