+++ TagCloudGenerator/TagCloudGenerator/GeneratorCore/CloudLayouters/ICloudLayouter.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagCloudGenerator.GeneratorCore.CloudLayouters
+{
+    public interface ICloudLayouter
+    {
+        Rectangle PutNextRectangle(Size rectangleSize);
+    }
+}
