+++ TagCloudGenerator/TagCloudGenerator/GeneratorCore/Tags/TagType.cs
@@ -1,0 +1,10 @@
+﻿namespace TagCloudGenerator.GeneratorCore.Tags
+{
+    public enum TagType
+    {
+        Central,
+        Large,
+        Medium,
+        Small
+    }
+}
