+++ TagCloudGenerator/TagCloudGenerator/Program.cs
@@ -1,14 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using TagCloudGenerator.CloudVocabularyPreprocessors;
-using TagCloudGenerator.Extensions;
-using TagCloudGenerator.TagClouds;
-using TagCloudGenerator.Tags;
-using TagCloudGenerator.UserInterfaces;
-using TagCloudGenerator.UserInterfaces.CmdClient;
-using TagCloudGenerator.UserInterfaces.CmdClient.CommandLineVerbs;
+﻿using TagCloudGenerator.Clients.CmdClient;
+using TagCloudGenerator.GeneratorCore;
+using TagCloudGenerator.GeneratorCore.CloudVocabularyPreprocessors;
 
 namespace TagCloudGenerator
 {
@@ -16,38 +8,22 @@ namespace TagCloudGenerator
     {
         private static void Main(string[] args)
         {
-            var cloudOptions = ArgumentParser.ReadCommandLineOptions(args);
-            var cloudContext = TagCloudOptionsParser.GetTegCloudContext(cloudOptions, CloudConstructor);
+            var cmdClient = new CommandLineClient(args);
+            var contextGenerator = new CloudContextGenerator(cmdClient);
+            var preprocessor = GetPreprocessor(contextGenerator);
+            var cloudGenerator = new CloudGenerator(contextGenerator, preprocessor);
 
-            cloudContext.TagCloudContent = ProcessVocabulary(cloudContext);
-
-            CreateTagCloudImage(cloudContext);
-
-            TagCloud<TagType> CloudConstructor(Color backgroundColor,
-                                               Dictionary<TagType, TagStyle> tagStyleByTagType) =>
-                cloudOptions is DoubleFontsCloud
-                    ? new CommonWordsCloud(backgroundColor, tagStyleByTagType)
-                    : (TagCloud<TagType>)new WebCloud(backgroundColor, tagStyleByTagType);
+            cloudGenerator.CreateTagCloudImage();
         }
 
-        private static IEnumerable<string> ProcessVocabulary(TagCloudContext cloudContext)
+        private static CloudVocabularyPreprocessor GetPreprocessor(CloudContextGenerator cloudContextGenerator)
         {
-            CloudVocabularyPreprocessor preprocessor = new ExcludingPreprocessors(null, cloudContext.ExcludedWords);
+            var cloudContext = cloudContextGenerator.GetTagCloudContext();
+
+            CloudVocabularyPreprocessor preprocessor = new ExcludingPreprocessors(null, cloudContext);
             preprocessor = new ToLowerPreprocessor(preprocessor);
 
-            return preprocessor.Process(cloudContext.TagCloudContent);
-        }
-
-        private static void CreateTagCloudImage(TagCloudContext cloudContext)
-        {
-            var shuffledContentStrings = cloudContext.TagCloudContent.Take(1)
-                .Concat(cloudContext.TagCloudContent.Skip(1).SequenceShuffle(new Random()))
-                .ToArray();
-
-            using var bitmap = cloudContext.Cloud.CreateBitmap(
-                shuffledContentStrings, cloudContext.CloudLayouter, cloudContext.ImageSize);
-
-            bitmap.Save(cloudContext.ImageName);
+            return preprocessor;
         }
     }
 }
