+++ TagCloudGenerator/TagCloudGenerator/TagCloudGenerator.csproj
@@ -48,31 +48,34 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CloudLayouters\ArchimedeanSpiral.cs" />
-    <Compile Include="CloudLayouters\CircularCloudLayouter.cs" />
-    <Compile Include="CloudLayouters\ICloudLayouter.cs" />
-    <Compile Include="CloudVocabularyPreprocessors\CloudVocabularyPreprocessor.cs" />
-    <Compile Include="CloudVocabularyPreprocessors\ExcludingPreprocessors.cs" />
-    <Compile Include="CloudVocabularyPreprocessors\ToLowerPreprocessor.cs" />
-    <Compile Include="Extensions\EnumerableExtensions.cs" />
-    <Compile Include="Extensions\RectangleExtensions.cs" />
+    <Compile Include="GeneratorCore\CloudContextGenerator.cs" />
+    <Compile Include="GeneratorCore\CloudGenerator.cs" />
+    <Compile Include="GeneratorCore\CloudLayouters\ArchimedeanSpiral.cs" />
+    <Compile Include="GeneratorCore\CloudLayouters\CircularCloudLayouter.cs" />
+    <Compile Include="GeneratorCore\CloudLayouters\ICloudLayouter.cs" />
+    <Compile Include="GeneratorCore\CloudVocabularyPreprocessors\CloudVocabularyPreprocessor.cs" />
+    <Compile Include="GeneratorCore\CloudVocabularyPreprocessors\ExcludingPreprocessors.cs" />
+    <Compile Include="GeneratorCore\CloudVocabularyPreprocessors\ToLowerPreprocessor.cs" />
+    <Compile Include="GeneratorCore\Extensions\EnumerableExtensions.cs" />
+    <Compile Include="GeneratorCore\Extensions\RectangleExtensions.cs" />
+    <Compile Include="GeneratorCore\TagCloudContext.cs" />
+    <Compile Include="GeneratorCore\TagClouds\CommonWordsCloud.cs" />
+    <Compile Include="GeneratorCore\TagClouds\ITagCloud.cs" />
+    <Compile Include="GeneratorCore\TagClouds\TagCloud.cs" />
+    <Compile Include="GeneratorCore\TagClouds\WebCloud.cs" />
+    <Compile Include="GeneratorCore\Tags\Tag.cs" />
+    <Compile Include="GeneratorCore\Tags\TagStyle.cs" />
+    <Compile Include="GeneratorCore\Tags\TagType.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TagCloudContext.cs" />
-    <Compile Include="TagClouds\CommonWordsCloud.cs" />
-    <Compile Include="TagClouds\TagCloud.cs" />
-    <Compile Include="TagClouds\WebCloud.cs" />
-    <Compile Include="Tags\Tag.cs" />
-    <Compile Include="Tags\TagStyle.cs" />
-    <Compile Include="Tags\TagType.cs" />
-    <Compile Include="UserInterfaces\CmdClient\ArgumentParser.cs" />
-    <Compile Include="UserInterfaces\TagCloudOptionsParser.cs" />
-    <Compile Include="UserInterfaces\CmdClient\CommandLineVerbs\DoubleFontsCloud.cs" />
-    <Compile Include="UserInterfaces\CmdClient\CommandLineVerbs\FourFontsCloud.cs" />
-    <Compile Include="UserInterfaces\CmdClient\CommandLineVerbs\TagCloudOptions.cs" />
-    <Compile Include="UserInterfaces\ITagCloudOptions.cs" />
-    <Compile Include="UserInterfaces\VocabularyParsers\CloudVocabularyParser.cs" />
-    <Compile Include="UserInterfaces\VocabularyParsers\TxtVocabularyParser.cs" />
+    <Compile Include="Clients\CmdClient\CommandLineClient.cs" />
+    <Compile Include="Clients\CmdClient\CommandLineVerbs\DoubleFontsCloud.cs" />
+    <Compile Include="Clients\CmdClient\CommandLineVerbs\FourFontsCloud.cs" />
+    <Compile Include="Clients\CmdClient\CommandLineVerbs\TagCloudOptions.cs" />
+    <Compile Include="Clients\IClient.cs" />
+    <Compile Include="Clients\ITagCloudOptions.cs" />
+    <Compile Include="Clients\VocabularyParsers\CloudVocabularyParser.cs" />
+    <Compile Include="Clients\VocabularyParsers\TxtVocabularyParser.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Images\CommonWordsTagCloud.png" />
