--- TagCloudGenerator/TagCloudGenerator/Tags/Tag.cs
@@ -1,18 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagCloudGenerator.Tags
-{
-    public struct Tag
-    {
-        public string Text { get; }
-        public TagStyle Style { get; }
-        public Rectangle TagBox { get; }
-
-        public Tag(string text, TagStyle style, Rectangle tagBox)
-        {
-            Text = text;
-            Style = style;
-            TagBox = tagBox;
-        }
-    }
-}
