--- TagCloudGenerator/TagCloudGenerator/Tags/TagType.cs
@@ -1,10 +1,0 @@
-﻿namespace TagCloudGenerator.Tags
-{
-    public enum TagType
-    {
-        Central,
-        Large,
-        Medium,
-        Small
-    }
-}
